Remove popular posts feature from blog
Production has hidden the blog sidebar using JavaScript, however the
"popular posts" feature in the sidebar is still being rendered out on
the server-side.  This functionality is causing excessive load on the
SQL server in some instances, bringing down the website.

This commit disables the feature, but leaves the functionality in place
if it is to be re-enabled later.

Issues: HYLT-248