Ensure "discounted subtotal" is properly processed every time
From debugging, it appeared that when the discounted subtotal validation
was ran, it either always saw the full price, or the completely
discounted price.

This fix updates the address amount so that, when it is validated before
being applied, it accurately grabs the discounted subtotal from the
address object.

Issues: HYLT-258 HYLT-234 HYLT-210 HYLT-219