Update MegaMenu module
This module had to be customized to support the new-style adminhtml
without compatibility mode (changes to config.xml and admin layout
handles).

As far as I could tell, there is no call out to megamenuprobackend.

I am unsure how or if this portion of Megamenu is used, and so it is not
as well tested as the rest.

Issues: HYLT-260